+++ 04_objektorientierung/13_solid/src/tpe/oo/metropolis/Mensch.java
@@ -2,9 +2,9 @@ package tpe.oo.metropolis;
 
 public abstract class Mensch extends Einwohner {
 
-    protected Mensch(String name, double einkommen) {
+    protected Mensch(String name, int einkommen) {
         super(name, einkommen);
-        
+
     }
 
 }
