+++ 04_objektorientierung/13_solid/src/tpe/oo/metropolis/Mutant.java
@@ -3,7 +3,7 @@ package tpe.oo.metropolis;
 public abstract class Mutant extends Einwohner {
 protected String mutation;
 
-protected Mutant(String name,double einkommen,String mutation){
+protected Mutant(String name,int einkommen,String mutation){
     super(name,einkommen);
     this.mutation=mutation;
 }
