+++ 04_objektorientierung/13_solid/src/tpe/oo/metropolis/Richter.java
@@ -2,12 +2,12 @@ package tpe.oo.metropolis;
 
 public class Richter extends Mensch {
     protected boolean korrupt;
-    protected Richter(String name, double einkommen,boolean korrupt) {
+    protected Richter(String name, int einkommen,boolean korrupt) {
         super(name, einkommen);
         this.korrupt=korrupt;
     }
-    
-    
+
+
     public void isKorrupt(boolean korrupt){
         this.korrupt=korrupt;
     }
